requirements.txt
flask
openai
requests
beautifulsoup4
langchain-community
faiss-cpu
telegram
langchain_openai
langchain-text-splitters
langgraph